requirements.txt
flask
gunicorn
pandas
openpyxl
chardet
python-dotenv
python-docx
